+++ alpha/apps/kaltura/modules/extwidget/actions/playManifestAction.class.php
@@ -711,9 +711,11 @@ class playManifestAction extends kalturaAction
 		if ($remoteFileSyncs)
 		{
 			$storageProfileIds = array_keys($remoteFileSyncs);
-			$storageProfiles = StorageProfilePeer::retrieveExternalByPartnerId(
+			$storageProfilesExternal = StorageProfilePeer::retrieveExternalByPartnerId(
 				$this->entry->getPartnerId(), 
 				$storageProfileIds);
+			$periodicProfiles = kStorageExporter::getPeriodicStorageProfiles($this->entry->getPartnerId());
+			$storageProfiles = array_merge($storageProfilesExternal, $periodicProfiles);
 
 			$activeStorageProfileIds = array();
 			foreach ($storageProfiles as $storageProfile)