+++ api_v3/lib/types/storageProfile/KalturaStorageProfile.php
@@ -197,11 +197,6 @@ class KalturaStorageProfile extends KalturaObject implements IFilterable
 	 */
 	public $regularPackagerUrl;
 
-	/**
-	 * @var bool
-	 */
-	public $exportPeriodically;
-
 	private static $map_between_objects = array
 	(
 		'id',
@@ -238,7 +233,6 @@ class KalturaStorageProfile extends KalturaObject implements IFilterable
 		'shouldExportThumbs',
 		'mappedPackagerUrl',
 		'regularPackagerUrl',
-		'exportPeriodically',
 	);
 	
 	/* (non-PHPdoc)